factoring reaction generation out from path finding
