+++ src/mrnet/network/reaction_network.py
@@ -13,7 +13,9 @@ from monty.json import MSONable
 from monty.serialization import dumpfn, loadfn
 from networkx.readwrite import json_graph
 
+from mrnet.network.reaction_generation import ReactionIterator
 from mrnet.core.mol_entry import MoleculeEntry
+from pymatgen.analysis.graphs import MoleculeGraph
 from mrnet.core.reactions import (
     ConcertedReaction,
     CoordinationBondChangeReaction,
@@ -36,16 +38,6 @@ __status__ = "Alpha"
 
 Mapping_Record_Dict = Dict[int, List[str]]
 RN_type = TypeVar("RN_type", bound="ReactionNetwork")
-
-
-test_dir = os.path.join(
-    os.path.dirname(__file__),
-    "..",
-    "..",
-    "..",
-    "test_files",
-    "reaction_network_files",
-)
 
 
 class ReactionPath(MSONable):
@@ -91,7 +83,6 @@ class ReactionPath(MSONable):
             "full_path": self.full_path,
         }
 
-    @property
     def as_dict(self) -> dict:
         """
             A method to convert ReactionPath objection into a dictionary
@@ -338,6 +329,12 @@ class ReactionPath(MSONable):
         }
         return class_instance
 
+    def __eq__(self, obj):
+        if type(self) == type(obj):
+            return self.as_dict() == obj.as_dict()
+        else:
+            return False
+
 
 Mapping_PR_Dict = Dict[int, Dict[int, ReactionPath]]
 
@@ -349,70 +346,21 @@ class ReactionNetwork(MSONable):
 
     def __init__(
         self,
-        electron_free_energy,
-        temperature,
-        solvent_dielectric,
-        solvent_refractive_index,
-        entries_dict,
-        entries_list,
-        graph,
-        reactions,
-        PRs,
-        PR_record,
-        min_cost,
-        num_starts,
-    ):
-        """
-        :param electron_free_energy: Electron free energy (in eV)
-        :param temperature: Temperature of the system, used for free energy
-            and rate constants (temperature given in K)
-        :param solvent_dielectric: dielectric constant of the solvent medium
-        :param solvent_refractive_index: refractive index of the solvent medium
-        :param entries_dict: dict of dicts of dicts of lists (d[formula][bonds][charge])
-        :param entries_list: list of unique entries in entries_dict
-        :param graph: nx.DiGraph representing connections in the network
-        :param reactions: list of Reaction objects
-        :param PRs: dict containing prerequisite information
-        :param PR_record: dict containing reaction prerequisites
-        :param min_cost: dict containing costs of entries in the network
-        :param num_starts: Number of starting molecules
-        """
-
-        self.electron_free_energy = electron_free_energy
-        self.temperature = temperature
-        self.solvent_dielectric = solvent_dielectric
-        self.solvent_refractive_index = solvent_refractive_index
-
-        self.entries = entries_dict
-        self.entries_list = entries_list
-
-        self.graph = graph
-        self.PR_record = PR_record
-        self.reactions = reactions
-
-        self.min_cost = min_cost
-        self.num_starts = num_starts
-
-        self.PRs = PRs
-        self.reachable_nodes = []
-        self.unsolvable_PRs = []
-        self.entry_ids = {e.entry_id for e in self.entries_list}
-        self.weight = None
-        self.Reactant_record = None
-        self.min_cost = {}
-        self.not_reachable_nodes = []
-        self.matrix = None
-        self.matrix_inverse = None
-
-    @classmethod
-    def from_input_entries(
-        cls,
-        input_entries,
+        reaction_iterator,
         electron_free_energy=-2.15,
         temperature=298.15,
         solvent_dielectric=18.5,
         solvent_refractive_index=1.415,
         replace_ind=True,
+        reaction_types: Union[Set, FrozenSet] = frozenset(
+            {
+                "RedoxReaction",
+                "IntramolSingleBondChangeReaction",
+                "IntermolecularReaction",
+                "CoordinationBondChangeReaction",
+            }
+        ),
+        determine_atom_mappings: bool = True,
     ):
         """
         Generate a ReactionNetwork from a set of MoleculeEntries.
@@ -430,102 +378,38 @@ class ReactionNetwork(MSONable):
         :return:
         """
 
-        entries = dict()
-        entries_list = list()
+        self.electron_free_energy = electron_free_energy
+        self.temperature = temperature
+        self.solvent_dielectric = solvent_dielectric
+        self.solvent_refractive_index = solvent_refractive_index
 
-        print(len(input_entries), "input entries")
+        self.entries = reaction_iterator.rn.entries
+        self.entries_list = reaction_iterator.rn.entries_list
 
-        # Filter out unconnected entries, aka those that contain distinctly
-        # separate molecules which are not connected via a bond
-        connected_entries = list()
-        for entry in input_entries:
-            if len(entry.molecule) > 1:
-                if nx.is_weakly_connected(entry.graph):
-                    connected_entries.append(entry)
-            else:
-                connected_entries.append(entry)
-        print(len(connected_entries), "connected entries")
+        self.graph = nx.DiGraph()
+        self.PR_record = dict()
+        self.reactions: List[Reaction] = list()
 
-        def get_formula(x):
-            return x.formula
+        self.PRs: dict = dict()
+        self.reachable_nodes: list = []
+        self.unsolvable_PRs: list = []
+        self.entry_ids = {e.entry_id for e in self.entries_list}
+        self.min_cost: dict = {}
+        self.not_reachable_nodes: list = []
 
-        def get_num_bonds(x):
-            return x.num_bonds
+        print("build() start", time.time())
 
-        def get_charge(x):
-            return x.charge
+        # Add molecule nodes
+        for entry in self.entries_list:
+            self.graph.add_node(entry.parameters["ind"], bipartite=0)
 
-        def get_free_energy(x):
-            return x.get_free_energy(temperature=temperature)
+        self.reactions = []
+        for reaction in reaction_iterator:
+            self.reactions.append(reaction)
+            self.add_reaction(reaction.graph_representation())
 
-        # Sort by formula
-        sorted_entries_0 = sorted(connected_entries, key=get_formula)
-        for k1, g1 in itertools.groupby(sorted_entries_0, get_formula):
-            sorted_entries_1 = sorted(list(g1), key=get_num_bonds)
-            entries[k1] = dict()
-            # Sort by number of bonds
-            for k2, g2 in itertools.groupby(sorted_entries_1, get_num_bonds):
-                sorted_entries_2 = sorted(list(g2), key=get_charge)
-                entries[k1][k2] = dict()
-                # Sort by charge
-                for k3, g3 in itertools.groupby(sorted_entries_2, get_charge):
-                    sorted_entries_3 = sorted(list(g3), key=get_free_energy)
-                    if len(sorted_entries_3) > 1:
-                        unique = list()
-                        for entry in sorted_entries_3:
-                            isomorphic_found = False
-                            # Sort by graph isomorphism, taking the isomorphic
-                            # entry with the lowest free energy
-                            for ii, Uentry in enumerate(unique):
-                                if entry.mol_graph.isomorphic_to(Uentry.mol_graph):
-                                    isomorphic_found = True
-                                    if (
-                                        entry.get_free_energy() is not None
-                                        and Uentry.get_free_energy() is not None
-                                    ):
-                                        if entry.get_free_energy(
-                                            temperature
-                                        ) < Uentry.get_free_energy(temperature):
-                                            unique[ii] = entry
-                                    elif entry.get_free_energy() is not None:
-                                        unique[ii] = entry
-                                    elif entry.energy < Uentry.energy:
-                                        unique[ii] = entry
-                                    break
-                            if not isomorphic_found:
-                                unique.append(entry)
-                        entries[k1][k2][k3] = unique
-                    else:
-                        entries[k1][k2][k3] = sorted_entries_3
-                    for entry in entries[k1][k2][k3]:
-                        entries_list.append(entry)
-
-        print(len(entries_list), "unique entries")
-        # Add entry indices
-        if replace_ind:
-            for ii, entry in enumerate(entries_list):
-                entry.parameters["ind"] = ii
-
-        entries_list = sorted(entries_list, key=lambda x: x.parameters["ind"])
-
-        graph = nx.DiGraph()
-
-        network = cls(
-            electron_free_energy,
-            temperature,
-            solvent_dielectric,
-            solvent_refractive_index,
-            entries,
-            entries_list,
-            graph,
-            list(),
-            dict(),
-            dict(),
-            dict(),
-            0,
-        )
-
-        return network
+        self.PR_record = self.build_PR_record()  # begin creating PR list
+        self.Reactant_record = self.build_reactant_record()  # begin creating rct list
 
     @staticmethod
     def softplus(free_energy: float) -> float:
@@ -549,89 +433,6 @@ class ReactionNetwork(MSONable):
         Method to determine edge weight using exponent(dG/kt) cost function
         """
         return rexp(free_energy)
-
-    def build(
-        self,
-        reaction_types: Union[Set, FrozenSet] = frozenset(
-            {
-                "RedoxReaction",
-                "IntramolSingleBondChangeReaction",
-                "IntermolecularReaction",
-                "CoordinationBondChangeReaction",
-            }
-        ),
-        determine_atom_mappings: bool = True,
-        build_matrix=False,
-    ) -> nx.DiGraph:
-        """
-            A method to build the reaction network graph
-
-        :param reaction_types (set/frozenset): set/frozenset of all the reactions
-            class to include while building the graph
-        :param determine_atom_mappings (bool): If True (default), create an atom
-            mapping between reactants and products in a given reaction
-        :return: nx.DiGraph
-        """
-
-        print("build() start", time.time())
-
-        # Add molecule nodes
-        for entry in self.entries_list:
-            self.graph.add_node(entry.parameters["ind"], bipartite=0)
-
-        reaction_classes = [load_class(str(self.__module__), s) for s in reaction_types]
-
-        all_reactions = list()
-
-        # Generate reactions
-        for r in reaction_classes:
-            reactions = r.generate(
-                self.entries, determine_atom_mappings=determine_atom_mappings
-            )  # review
-            all_reactions.append(reactions)
-
-        all_reactions = [i for i in all_reactions if i]
-        self.reactions = list(itertools.chain.from_iterable(all_reactions))
-
-        redox_c = 0
-        inter_c = 0
-        intra_c = 0
-        coord_c = 0
-
-        for ii, r in enumerate(self.reactions):
-            r.parameters["ind"] = ii
-            if r.__class__.__name__ == "RedoxReaction":
-                redox_c += 1
-                r.electron_free_energy = self.electron_free_energy
-                r.set_free_energy()
-                r.set_rate_constant()
-            elif r.__class__.__name__ == "IntramolSingleBondChangeReaction":
-                intra_c += 1
-            elif r.__class__.__name__ == "IntermolecularReaction":
-                inter_c += 1
-            elif r.__class__.__name__ == "CoordinationBondChangeReaction":
-                coord_c += 1
-            self.add_reaction(r.graph_representation())  # add graph element here
-
-        print(
-            "redox: ",
-            redox_c,
-            "inter: ",
-            inter_c,
-            "intra: ",
-            intra_c,
-            "coord: ",
-            coord_c,
-        )
-        self.PR_record = self.build_PR_record()  # begin creating PR list
-        self.Reactant_record = self.build_reactant_record()  # begin creating rct list
-
-        if build_matrix:
-            self.build_matrix()
-
-        print("build() end", time.time())
-
-        return self.graph
 
     def add_reaction(self, graph_representation: nx.DiGraph):
         """
@@ -722,14 +523,6 @@ class ReactionNetwork(MSONable):
         self.num_starts = len(starts)
         self.PR_byproducts = {}  # type: Dict[int, Dict[str, int]]
 
-        if len(self.graph.nodes) == 0:
-            self.build()  # actually construct the graph
-        if self.PR_record is None:
-            self.PR_record = self.build_PR_record()  # get a dict of PRs
-        if self.Reactant_record is None:
-            self.Reactant_record = (
-                self.build_reactant_record()
-            )  # get a dict of non-PR reactants
         orig_graph = copy.deepcopy(self.graph)
 
         for start in starts:  # all the molecular nodes
@@ -954,7 +747,7 @@ class ReactionNetwork(MSONable):
                                     int(paths[node][ii - 1])
                                 )  # source reactant
                                 rcts.remove(
-                                    paths[node][ii - 1]
+                                    str(paths[node][ii - 1])
                                 )  # remove "reactant" reactant
                                 assert len(rcts) == 1
                                 PR.append(int(rcts[0]))
@@ -1126,7 +919,7 @@ class ReactionNetwork(MSONable):
         self,
         min_cost: Dict[int, float],
         orig_graph: nx.DiGraph,
-    ) -> Dict[Tuple[int, str], Dict[str, float]]:
+    ):
         """
             A method to update the ReactionNetwork.graph edge weights based on
             the new cost of solving PRs
@@ -1368,60 +1161,6 @@ class ReactionNetwork(MSONable):
         return self.PRs, paths, top_path_list
 
     @staticmethod
-    def mols_w_cuttoff(RN_pr_solved, cutoff=0, build_pruned_network=True):
-        """
-        A method to identify molecules reached by dG <= cutoff
-
-        :param RN_pr_solved: instance of reaction network
-        :param: cutoff: dG value
-        :param: build_pruned_network: if true a network with pruned entries will be build
-        :return: mols_to_keep: list of molecule nodes that can be reached by dG <= cutoff
-        :return: pruned_entries_list: list of MoleculeEntry of molecules that can be reached by dG <= cutoff
-        """
-
-        pruned_PRs = {}
-        for PR_node in RN_pr_solved.PR_byproducts:
-            if (
-                RN_pr_solved.PRs[PR_node] != {}
-                and RN_pr_solved.PR_byproducts[PR_node] != {}
-            ):
-                min_start = RN_pr_solved.PR_byproducts[PR_node]["start"]
-                if (
-                    RN_pr_solved.PRs[PR_node][min_start].overall_free_energy_change
-                    <= cutoff
-                ):
-                    pruned_PRs[PR_node] = {}
-                    pruned_PRs[PR_node][min_start] = RN_pr_solved.PRs[PR_node][
-                        min_start
-                    ]
-
-        nodes_to_keep = []
-        for PR_node in pruned_PRs:
-            for start in pruned_PRs[PR_node]:
-                nodes_to_keep = nodes_to_keep + pruned_PRs[PR_node][start].full_path
-
-        nodes_to_keep = list(dict.fromkeys(nodes_to_keep))
-        mols_to_keep = []
-        for node in nodes_to_keep:
-            if isinstance(node, int):
-                mols_to_keep.append(node)
-        mols_to_keep.sort()
-
-        pruned_entries_list = []
-        for entry in RN_pr_solved.entries_list:
-            if entry.parameters["ind"] in mols_to_keep:
-                pruned_entries_list.append(entry)
-
-        if build_pruned_network:
-            pruned_network_build = ReactionNetwork.from_input_entries(
-                pruned_entries_list, replace_ind=False
-            )
-            pruned_network_build.build()
-            return mols_to_keep, pruned_entries_list, pruned_network_build
-        else:
-            return mols_to_keep, pruned_entries_list
-
-    @staticmethod
     def parse_reaction_node(node: str):
         """
         A method to identify reactants, PR, and prodcuts from a given reaction node string.
@@ -1451,301 +1190,6 @@ class ReactionNetwork(MSONable):
             + "+".join(list(map(str, combined_products)))
         )
         return node_str
-
-    @staticmethod
-    def concerted_reaction_filter(in_reaction_node, out_reaction_node):
-        """
-        A method to identify a valid concerted reaction based on stiochomtery of maximum of 2 reactants and products
-        :param in_reaction_node: incoming reaction node "6,2+7"
-        :param out_reaction_node: outgoing reaction node "2+1,3"
-        :return: r: combined reaction's reactant and product [[1,6],[3,7]], r_node: combined reaction's reactant and
-        product as well as the two reaction nodes - ex [[1,6],[3,7], "6,2+7","2+1,3"]
-        """
-        r = None
-        r_node = None
-        unique_reactions = []
-        (
-            in_reactants,
-            in_products,
-        ) = ReactionNetwork.parse_reaction_node(in_reaction_node)
-        (out_reactants, out_products) = ReactionNetwork.parse_reaction_node(
-            out_reaction_node
-        )
-        combined_reactants = in_reactants + out_reactants
-        combined_products = in_products + out_products
-        combined_reactants.sort()
-        combined_products.sort()
-        inter = set(combined_reactants).intersection(set(combined_products))
-        for i in inter:
-            combined_reactants.remove(i)
-            combined_products.remove(i)
-        inter = set(combined_reactants).intersection(set(combined_products))
-        for i in inter:
-            combined_reactants.remove(i)
-            combined_products.remove(i)
-        if 0 < len(combined_reactants) <= 2 and 0 < len(combined_products) <= 2:
-            r = [combined_reactants, combined_products]
-            unique_reactions.append(r)
-            r_node = [
-                combined_reactants,
-                combined_products,
-                [in_reaction_node, out_reaction_node],
-            ]
-        return r, r_node
-
-    def build_matrix(self) -> Dict[int, Dict[int, List[Tuple]]]:
-        """
-        A method to build a spare adjacency matrix using dictionaries.
-        :return: nested dictionary {r1:{c1:[],c2:[]}, r2:{c1:[],c2:[]}}
-        """
-        self.matrix = {}
-        for i in range(len(self.entries_list)):
-            self.matrix[i] = {}
-        for node in self.graph.nodes:
-            if isinstance(node, str):
-                if "electron" not in self.graph.nodes[node]["rxn_type"]:
-                    in_node = list(self.graph.predecessors(node))
-                    out_nodes = list(self.graph.successors(node))
-                    edges = []
-                    for u in in_node:
-                        for v in out_nodes:
-                            edges.append((u, v))
-                    for e in edges:
-                        if e[1] not in self.matrix[e[0]].keys():
-                            self.matrix[e[0]][e[1]] = [
-                                (node, self.graph.nodes[node]["free_energy"], "e")
-                            ]
-                        else:
-                            self.matrix[e[0]][e[1]].append(
-                                (node, self.graph.nodes[node]["free_energy"], "e")
-                            )
-        self.matrix_inverse = {}
-        for i in range(len(self.matrix)):
-            self.matrix_inverse[i] = {}
-            for k, v in self.matrix.items():
-                if i in v.keys():
-                    self.matrix_inverse[i][k] = v[i]
-
-        return self.matrix
-
-    @staticmethod
-    def identify_concerted_rxns_via_intermediates(
-        RN,
-        mols_to_keep=None,
-        single_elem_interm_ignore=["C1", "H1", "O1", "Li1", "P1", "F1"],
-        update_matrix=False,
-    ):
-        """
-            A method to identify concerted reactions via high enery intermediate molecules
-        :param RN: Reaction network built
-        :param mols_to_keep: List of pruned molecules, if not running then a list of all molecule nodes in the
-        RN
-        :param single_elem_interm_ignore: single_elem_interm_ignore: List of formula of high energy
-        intermediates to ignore
-        :param update_matrix: If true, self.matrix2 will be updated with iteration 1 concerted reaction
-        :return: list of unique reactions, list of reactions and its incoming and outgoing reaction nodes
-        """
-
-        print("identify_concerted_rxns_via_intermediates start", time.time())
-        if mols_to_keep is None:
-            mols_to_keep = list(range(0, len(RN.entries_list)))
-        if update_matrix:
-            RN.matrix2 = None
-        reactions = []
-        unique_reactions = []
-        for entry in RN.entries_list:
-            (
-                unique_rxns,
-                rxns_with_nodes,
-            ) = RN.identify_concerted_rxns_for_specific_intermediate(
-                entry, RN, mols_to_keep, single_elem_interm_ignore, update_matrix
-            )
-            unique_reactions.append(unique_rxns)
-            reactions.append(rxns_with_nodes)
-        all_unique_reactions = reduce(operator.concat, unique_reactions)
-        all_unique_reactions = list(
-            map(
-                lambda y: literal_eval(y),
-                set(map(lambda x: repr(x), all_unique_reactions)),
-            )
-        )
-
-        print("total number of unique concerted reactions:", len(all_unique_reactions))
-        print("identify_concerted_rxns_via_intermediates end", time.time())
-        return all_unique_reactions, reactions
-
-    @staticmethod
-    def add_reactions_to_matrix(matrix, reaction):
-        """
-        A method to add new concerted reactions to the matrix which is already built from elemetary reactions.
-        :param matrix: matrix which is already built from elemetary reactions, self.matrix
-        :param reaction: concerted reaction to add to the matrix, (1,2], [3,4], total_dG)
-        :return: matrix updated with the reaction
-        """
-        nstr = ReactionNetwork.generate_node_string(reaction[0], reaction[1])
-        for reac in reaction[0]:
-            for prod in reaction[1]:
-                if prod not in matrix[reac].keys():
-                    matrix[reac][prod] = [(nstr, reaction[2], "c")]
-                else:
-                    matrix[reac][prod].append((nstr, reaction[2], "c"))
-        return matrix
-
-    @staticmethod
-    def identify_concerted_rxns_for_specific_intermediate(
-        entry: MoleculeEntry,
-        RN: RN_type,
-        mols_to_keep=None,
-        single_elem_interm_ignore=["C1", "H1", "O1", "Li1", "P1", "F1"],
-        update_matrix=False,
-    ):
-
-        """
-            A method to identify concerted reactions via specific high enery intermediate molecule
-        :param entry: MoleculeEntry to act as high energy intermediate
-        :param RN: Reaction network built
-        :param mols_to_keep: List of pruned molecules, if not running then a list of all molecule nodes
-        in the RN_pr_solved
-        :param single_elem_interm_ignore: single_elem_interm_ignore: List of formula of high energy
-        intermediates to ignore
-        :return: list of reactions
-        """
-
-        rxn_from_filer_iter1 = []
-        rxn_from_filer_iter1_nodes = []
-        entry_ind = entry.parameters["ind"]  # type: int
-
-        if mols_to_keep is None:
-            mols_to_keep = list(range(0, len(RN.entries_list)))
-        not_wanted_formula = single_elem_interm_ignore
-
-        if (
-            entry.formula not in not_wanted_formula
-            and entry.parameters["ind"] in mols_to_keep
-        ):
-
-            if RN.matrix is None:
-                RN.build_matrix()
-            if update_matrix:
-                RN.matrix2 = copy.deepcopy(RN.matrix)
-
-            row = RN.matrix[entry_ind]  # type: ignore
-            col = RN.matrix_inverse[entry_ind]
-
-            for kr, vr in row.items():
-                for kc, vc in col.items():
-                    if kr != kc:
-                        for s2 in vr:
-                            for e2 in vc:
-                                incoming_reaction_dG = e2[1]
-                                total_dG = s2[1] + e2[1]
-                                if incoming_reaction_dG > 0 and total_dG < 0:
-                                    (
-                                        rxn1,
-                                        rxn1_nodes,
-                                    ) = ReactionNetwork.concerted_reaction_filter(
-                                        e2[0], s2[0]
-                                    )
-                                    if rxn1 is not None:
-                                        rxn_from_filer_iter1.append(rxn1)
-                                        rxn_from_filer_iter1_nodes.append(rxn1_nodes)
-                                        if update_matrix:
-                                            reaction = (rxn1[0], rxn1[1], total_dG)
-                                            ReactionNetwork.add_reactions_to_matrix(
-                                                RN.matrix2, reaction
-                                            )
-
-        return rxn_from_filer_iter1, rxn_from_filer_iter1_nodes
-
-    @staticmethod
-    def add_concerted_rxns(RN, reactions):
-        """
-            A method to add concerted reactions (obtained from identify_concerted_rxns_via_intermediates() method)to
-            the ReactonNetwork
-        :param RN: build Reaction Network
-        :param reactions: list of reactions obtained from identify_concerted_rxns_via_intermediates() method
-        :return: reaction network with concerted reactions added
-        """
-
-        print("add_concerted_rxns start", time.time())
-        c1 = 0
-        c2 = 0
-        c3 = 0
-        c4 = 0
-        mol_id_to_mol_entry_dict = {}
-        for i in RN.entries_list:
-            mol_id_to_mol_entry_dict[int(i.parameters["ind"])] = i
-        for reaction in reactions:
-            if len(reaction[0]) == 1 and len(reaction[1]) == 1:
-                assert int(reaction[0][0]) in RN.graph.nodes
-                assert int(reaction[1][0]) in RN.graph.nodes
-                reactants = mol_id_to_mol_entry_dict[int(reaction[0][0])]
-                products = mol_id_to_mol_entry_dict[int(reaction[1][0])]
-                cr = ConcertedReaction([reactants], [products])
-                cr.electron_free_energy = RN.electron_free_energy
-                g = cr.graph_representation()
-                for node in list(g.nodes):
-                    if not isinstance(node, int) and g.nodes[node]["free_energy"] > 0:
-                        g.remove_node(node)
-                RN.add_reaction(g)
-                c1 = c1 + 1
-
-            elif len(reaction[0]) == 1 and len(reaction[1]) == 2:
-                assert int(reaction[0][0]) in RN.graph.nodes
-                assert int(reaction[1][0]) in RN.graph.nodes
-                assert int(reaction[1][1]) in RN.graph.nodes
-                reactant_0 = mol_id_to_mol_entry_dict[int(reaction[0][0])]
-                product_0 = mol_id_to_mol_entry_dict[int(reaction[1][0])]
-                product_1 = mol_id_to_mol_entry_dict[int(reaction[1][1])]
-                cr = ConcertedReaction([reactant_0], [product_0, product_1])
-                cr.electron_free_energy = RN.electron_free_energy
-                g = cr.graph_representation()
-                for node in list(g.nodes):
-                    if not isinstance(node, int) and g.nodes[node]["free_energy"] > 0:
-                        g.remove_node(node)
-
-                RN.add_reaction(g)
-                c2 = c2 + 1
-            elif len(reaction[0]) == 2 and len(reaction[1]) == 1:
-                assert int(reaction[0][0]) in RN.graph.nodes
-                assert int(reaction[0][1]) in RN.graph.nodes
-                assert int(reaction[1][0]) in RN.graph.nodes
-                reactant_0 = mol_id_to_mol_entry_dict[int(reaction[0][0])]
-                PR = mol_id_to_mol_entry_dict[int(reaction[0][1])]
-                product_0 = mol_id_to_mol_entry_dict[int(reaction[1][0])]
-                cr = ConcertedReaction([reactant_0, PR], [product_0])
-                cr.electron_free_energy = RN.electron_free_energy
-                g = cr.graph_representation()
-                for node in list(g.nodes):
-                    if not isinstance(node, int) and g.nodes[node]["free_energy"] > 0:
-                        g.remove_node(node)
-
-                RN.add_reaction(g)
-                c3 = c3 + 1
-            elif len(reaction[0]) == 2 and len(reaction[1]) == 2:
-                assert int(reaction[0][0]) in RN.graph.nodes
-                assert int(reaction[0][1]) in RN.graph.nodes
-                assert int(reaction[1][0]) in RN.graph.nodes
-                assert int(reaction[1][1]) in RN.graph.nodes
-                reactant_0 = mol_id_to_mol_entry_dict[int(reaction[0][0])]
-                PR = mol_id_to_mol_entry_dict[int(reaction[0][1])]
-                product_0 = mol_id_to_mol_entry_dict[int(reaction[1][0])]
-                product_1 = mol_id_to_mol_entry_dict[int(reaction[1][1])]
-                cr = ConcertedReaction([reactant_0, PR], [product_0, product_1])
-                cr.electron_free_energy = RN.electron_free_energy
-                g = cr.graph_representation()
-                for node in list(g.nodes):
-                    if not isinstance(node, int) and g.nodes[node]["free_energy"] > 0:
-                        g.remove_node(node)
-
-                RN.add_reaction(g)
-                c4 = c4 + 1
-        total_num_concerted = c1 + c2 + c3 + c4
-        RN.PR_record = RN.build_PR_record()
-        RN.Reactant_record = RN.build_reactant_record()
-        print("number of concerted reactions added", total_num_concerted)
-        print("add_concerted_rxns end", time.time())
-        return RN
 
     def as_dict(self) -> dict:
         entries = dict()  # type: Dict[str, Dict[int, Dict[int, List[Dict[str, Any]]]]]
@@ -1823,95 +1267,20 @@ class ReactionNetwork(MSONable):
             d["num_starts"],
         )
 
-    def generate_pre_find_path_files(
-        self, PRs, cost_from_start, old_solved_PRs, min_cost
-    ):
-        pickle_in = open(
-            os.path.join(test_dir, "unittest_RN_pr_ii_4_before_find_path_cost.pkl"),
-            "wb",
-        )
-        pickle.dump(self, pickle_in)
-        with open(
-            os.path.join(test_dir, "unittest_find_path_cost_PRs_IN.pkl"),
-            "wb",
-        ) as handle:
-            pickle.dump(PRs, handle, protocol=4)
-        dumpfn(
-            cost_from_start,
-            os.path.join(test_dir, "unittest_find_path_cost_cost_from_start_IN.json"),
-        )
-        dumpfn(
-            old_solved_PRs,
-            os.path.join(test_dir, "unittest_find_path_cost_old_solved_prs_IN.json"),
-        )
-        dumpfn(
-            min_cost,
-            os.path.join(test_dir, "unittest_find_path_cost_min_cost_IN.json"),
-        )
 
-    def generate_pre_id_solved_PRs_files(self, PRs, cost_from_start, solved_PRs):
-        pickle_in = open(
-            os.path.join(
-                test_dir,
-                "unittest_RN_pr_ii_4_before_identify_solved_PRs.pkl",
-            ),
-            "wb",
-        )
-        pickle.dump(self, pickle_in)
-        with open(
-            os.path.join(test_dir, "unittest_find_path_cost_PRs_IN.pkl"),
-            "wb",
-        ) as handle:
-            pickle.dump(PRs, handle, protocol=4)
-        dumpfn(
-            cost_from_start,
-            os.path.join(
-                test_dir,
-                "unittest_identify_solved_PRs_cost_from_start_IN.json",
-            ),
-        )
-        dumpfn(
-            solved_PRs,
-            os.path.join(test_dir, "unittest_identify_solved_PRs_solved_PRs_IN.json"),
-        )
+def path_finding_wrapper(
+    mol_list: List[MoleculeEntry], init_mols: List[MoleculeEntry], target: MoleculeEntry
+):
+    ri = ReactionIterator(mol_list)
+    rn = ReactionNetwork(ri)
 
-    def generate_characterize_path_files(self, old_solved_PRs, dist_and_path):
-        pickle_in = open(
-            os.path.join(
-                test_dir,
-                "unittest_RN_before_characterize_path.pkl",
-            ),
-            "wb",
-        )
-        pickle.dump(self, pickle_in)
-        pickle_in = open(
-            os.path.join(test_dir, "unittest_characterize_path_PRs_IN.pkl"),
-            "wb",
-        )
-        pickle.dump(old_solved_PRs, pickle_in)
-        dumpfn(
-            dist_and_path,
-            os.path.join(
-                test_dir,
-                "unittest_characterize_path_path_IN.json",
-            ),
-        )
+    initial_inds = [e.parameters["ind"] for e in init_mols]
 
-    def generate_pre_update_eweights_files(self, min_cost):
-        pickle_in = open(
-            os.path.join(
-                test_dir,
-                "unittest_RN_pr_ii_4_before_update_edge_weights.pkl",
-            ),
-            "wb",
-        )
-        pickle.dump(self, pickle_in)
-        pickle_in = open(
-            os.path.join(test_dir, "unittest_update_edge_weights_orig_graph_IN.pkl"),
-            "wb",
-        )
-        pickle.dump(self.graph, pickle_in)
-        dumpfn(
-            min_cost,
-            os.path.join(test_dir, "unittest_update_edge_weights_min_cost_IN.json"),
-        )
+    rn.solve_prerequisites(initial_inds, "softplus")
+
+    # set initial conditions
+    PRs, paths, top_path_list = rn.find_paths(
+        initial_inds, target.parameters["ind"], weight="softplus", num_paths=20
+    )
+    # return shortest paths to every mol
+    return PRs
