+++ src/mrnet/stochastic/serialize.py
@@ -12,7 +12,7 @@ from pymatgen.analysis.local_env import OpenBabelNN
 from pymatgen.analysis.fragmenter import metal_edge_extender
 
 from mrnet.network.reaction_network import ReactionNetwork
-from mrnet.network.reaction_generation import ReactionGenerator
+from mrnet.network.reaction_generation import ReactionIterator
 from mrnet.core.mol_entry import MoleculeEntry
 from mrnet.utils.constants import ROOM_TEMP, PLANCK, KB
 
@@ -140,7 +140,7 @@ class SerializeNetwork:
     def __init__(
         self,
         folder: str,
-        reaction_generator: ReactionGenerator,
+        reaction_generator: ReactionIterator,
         shard_size: int = 1000000,
         commit_barrier: int = 10000,
         temperature=ROOM_TEMP,
@@ -252,12 +252,11 @@ class SerializeNetwork:
 
     def serialize(self):
 
-        for (
-            reactants,
-            products,
-            forward_free_energy,
-            backward_free_energy,
-        ) in self.reaction_generator:
+        for reaction in self.reaction_generator:
+            reactants = tuple([int(r) for r in reaction.reactant_indices])
+            products = tuple([int(r) for r in reaction.product_indices])
+            forward_free_energy = reaction.free_energy_A
+            backward_free_energy = reaction.free_energy_B
 
             forward_reaction_string = "".join(
                 [
