+++ test_files/reaction_network_files/ronalds_concerteds.json
@@ -1,0 +1,1 @@
+[[[0], [3, 37]], [[0], [4, 36]], [[1], [8, 45]], [[2], [9, 45]], [[13], [10, 3]], [[14], [3, 10]], [[15], [4, 10]], [[16], [4, 10]], [[21], [10, 6]], [[25], [13, 40]], [[25], [15, 39]], [[26], [15, 40]], [[27], [15, 38]], [[27], [22, 36]], [[27], [25, 45]], [[28], [22, 37]], [[28], [26, 45]], [[31], [3, 15]], [[31], [4, 13]], [[32], [4, 15]], [[33], [16, 4]], [[33], [4, 15]], [[34], [6, 15]], [[34], [21, 4]], [[38], [45, 39]], [[41], [45, 43]], [[41], [38, 36]], [[42], [45, 44]], [[42], [38, 37]], [[43], [39, 36]], [[44], [39, 37]], [[44], [40, 36]], [[6], [3, 45]], [[18], [23, 45]], [[19], [24, 45]], [[21], [13, 45]], [[21], [14, 45]], [[32], [29, 45]], [[33], [29, 45]], [[34], [31, 45]], [[35], [32, 45]], [[35], [33, 45]], [[1], [2]], [[3], [4]], [[8], [9]], [[13], [15]], [[14], [16]], [[18], [19]], [[23], [24]], [[25], [26]], [[27], [28]], [[31], [32]], [[34], [35]], [[36], [37]], [[39], [40]], [[41], [42]], [[43], [44]], [[16], [18]], [[15], [18]], [[6, 37], [0, 45]], [[15, 36], [0, 10]], [[16, 36], [0, 10]], [[31, 36], [0, 13]], [[32, 36], [0, 15]], [[33, 36], [0, 16]], [[33, 36], [0, 15]], [[16], [15]], [[13, 15], [10, 31]], [[13, 16], [10, 31]], [[15, 15], [10, 32]], [[15, 16], [10, 32]], [[15, 31], [13, 32]], [[15, 33], [16, 32]], [[33], [32]], [[15, 16], [10, 33]], [[15, 15], [10, 33]], [[16, 16], [10, 33]], [[15, 16], [10, 33]], [[16, 31], [13, 33]], [[15, 31], [13, 33]], [[0, 21], [34, 36]], [[15, 21], [10, 34]], [[16, 21], [10, 34]], [[21, 31], [13, 34]], [[21, 32], [15, 34]], [[21, 33], [16, 34]], [[21, 33], [15, 34]], [[16], [15]], [[31, 40], [4, 25]], [[25, 40], [26, 39]], [[32, 40], [4, 26]], [[33, 40], [4, 26]], [[4, 25], [32, 39]], [[33], [32]], [[4, 25], [33, 39]], [[16], [15]], [[13, 45], [6, 10]], [[14, 45], [6, 10]], [[13, 45], [6, 10]], [[14, 45], [6, 10]], [[13, 45], [6, 10]], [[14, 45], [6, 10]], [[13, 45], [6, 10]], [[14, 45], [6, 10]], [[14], [13]], [[14], [13]], [[14], [13]], [[27, 37], [28, 36]], [[33], [32]], [[4, 27], [0, 22]], [[4, 41], [0, 38]], [[4, 43], [0, 39]], [[27, 38], [22, 41]], [[27, 39], [22, 43]], [[39, 41], [38, 43]], [[0, 40], [4, 44]], [[27, 40], [22, 44]], [[40, 41], [38, 44]], [[40, 43], [39, 44]], [[3, 28], [0, 22]], [[3, 42], [0, 38]], [[28, 38], [22, 42]], [[0, 39], [3, 44]], [[28, 39], [22, 44]], [[39, 42], [38, 44]], [[37, 41], [36, 42]], [[25, 45], [15, 38]], [[43, 45], [36, 38]], [[25, 36], [15, 43]], [[25, 37], [15, 44]], [[37, 38], [44, 45]], [[37, 43], [36, 44]], [[15, 25], [13, 26]], [[25, 36], [13, 44]], [[26, 36], [15, 44]], [[2, 8], [1, 9]], [[8, 19], [1, 24]], [[8, 27], [1, 25]], [[8, 28], [1, 26]], [[8, 35], [1, 32]], [[8, 35], [1, 33]], [[8, 42], [1, 44]], [[2, 25], [9, 27]], [[19, 25], [24, 27]], [[25, 28], [26, 27]], [[25, 35], [27, 32]], [[25, 35], [27, 33]], [[25, 42], [27, 44]], [[2, 26], [9, 28]], [[19, 26], [24, 28]], [[26, 42], [28, 44]], [[1, 39], [8, 38]], [[2, 39], [9, 38]], [[6, 39], [3, 38]], [[18, 39], [23, 38]], [[19, 39], [24, 38]], [[27, 39], [25, 38]], [[28, 39], [26, 38]], [[32, 39], [29, 38]], [[33, 39], [29, 38]], [[34, 39], [31, 38]], [[35, 39], [32, 38]], [[35, 39], [33, 38]], [[39, 41], [38, 43]], [[39, 42], [38, 44]], [[1, 43], [8, 41]], [[2, 43], [9, 41]], [[18, 43], [23, 41]], [[19, 43], [24, 41]], [[27, 43], [25, 41]], [[28, 43], [26, 41]], [[32, 43], [29, 41]], [[33, 43], [29, 41]], [[34, 43], [31, 41]], [[35, 43], [32, 41]], [[35, 43], [33, 41]], [[42, 43], [41, 44]], [[2, 44], [9, 42]], [[1, 3], [6, 8]], [[2, 3], [6, 9]], [[3, 18], [6, 23]], [[3, 19], [6, 24]], [[3, 27], [6, 25]], [[3, 28], [6, 26]], [[3, 32], [6, 29]], [[3, 33], [6, 29]], [[3, 34], [6, 31]], [[3, 35], [6, 32]], [[3, 35], [6, 33]], [[3, 41], [6, 43]], [[3, 42], [6, 44]], [[1, 23], [8, 18]], [[2, 23], [9, 18]], [[19, 23], [18, 24]], [[23, 27], [18, 25]], [[23, 28], [18, 26]], [[23, 33], [18, 29]], [[23, 35], [18, 32]], [[23, 35], [18, 33]], [[23, 42], [18, 44]], [[2, 24], [9, 19]], [[24, 42], [19, 44]], [[1, 29], [8, 32]], [[2, 29], [9, 32]], [[18, 29], [23, 32]], [[19, 29], [24, 32]], [[27, 29], [25, 32]], [[28, 29], [26, 32]], [[33], [32]], [[29, 35], [32, 32]], [[29, 35], [32, 33]], [[29, 42], [32, 44]], [[1, 29], [8, 33]], [[2, 29], [9, 33]], [[19, 29], [24, 33]], [[27, 29], [25, 33]], [[28, 29], [26, 33]], [[29, 35], [32, 33]], [[29, 35], [33, 33]], [[29, 42], [33, 44]], [[1, 31], [8, 34]], [[2, 31], [9, 34]], [[18, 31], [23, 34]], [[19, 31], [24, 34]], [[27, 31], [25, 34]], [[28, 31], [26, 34]], [[31, 32], [29, 34]], [[31, 33], [29, 34]], [[31, 35], [32, 34]], [[31, 35], [33, 34]], [[31, 42], [34, 44]], [[2, 32], [9, 35]], [[2, 33], [9, 35]], [[19, 32], [24, 35]], [[19, 33], [24, 35]], [[28, 32], [26, 35]], [[28, 33], [26, 35]], [[32, 42], [35, 44]], [[33, 42], [35, 44]]]
