+++ tests/stochastic/test_rnmc.py
@@ -10,7 +10,7 @@ from scipy.constants import N_A
 from monty.serialization import loadfn, dumpfn
 from pymatgen.util.testing import PymatgenTest
 
-from mrnet.network.reaction_generation import ReactionGenerator
+from mrnet.network.reaction_generation import ReactionIterator
 from mrnet.stochastic.serialize import (
     SerializeNetwork,
     serialize_simulation_parameters,
@@ -36,6 +36,36 @@ test_dir = os.path.join(
     "test_files",
     "reaction_network_files",
 )
+
+
+class TestReactionGenerator(PymatgenTest):
+    def test_reaction_generator(self):
+
+        molecule_entries = loadfn(os.path.join(test_dir, "ronalds_MoleculeEntry.json"))
+        reaction_generator = ReactionIterator(
+            molecule_entries, single_elem_interm_ignore=[]
+        )
+        reactions = []
+
+        for reaction in reaction_generator:
+            reactions.append(
+                (
+                    tuple([int(r) for r in reaction.reactant_indices]),
+                    tuple([int(r) for r in reaction.product_indices]),
+                )
+            )
+
+        result = frozenset(reactions)
+
+        # ronalds concerteds is a json dump of reactions since we can't serialize frozensets to json
+        ronalds_concerteds_lists = loadfn(
+            os.path.join(test_dir, "ronalds_concerteds.json")
+        )
+        ronalds_concerteds = frozenset(
+            [tuple([tuple(x[0]), tuple(x[1])]) for x in ronalds_concerteds_lists]
+        )
+
+        assert result == ronalds_concerteds
 
 
 class RNMC(PymatgenTest):
@@ -66,7 +96,7 @@ class RNMC(PymatgenTest):
         initial_state_data_1 = [(li_plus_mol_entry, 300), (ec_mol_entry, 30)]
         initial_state_data_2 = [(li_plus_mol_entry, 30), (ec_mol_entry, 300)]
 
-        reaction_generator = ReactionGenerator(molecule_entries)
+        reaction_generator = ReactionIterator(molecule_entries)
 
         # for large networks, you want to use shard_size=2000000
         SerializeNetwork(network_folder_1, reaction_generator, shard_size=100)
